+++ src/textual/widgets/_text_area.py
@@ -240,7 +240,6 @@ TextArea {
         ),
         Binding("ctrl+z", "undo", "Undo", show=False),
         Binding("ctrl+y", "redo", "Redo", show=False),
-        Binding("ctrl+c", "copy_selection", "Copy selected text", show=False),
     ]
     """
     | Key(s)                 | Description                                  |
@@ -274,6 +273,9 @@ TextArea {
     | f7                     | Select all text in the document.             |
     | ctrl+z                 | Undo.                                        |
     | ctrl+y                 | Redo.                                        |
+    | ctrl+x                 | Cut selection or line if no selection.       |
+    | ctrl+c                 | Copy selection to clipboard.                 |
+    | ctrl+v                 | Paste from clipboard.                        |
     """
 
     language: Reactive[str | None] = reactive(None, always_update=True, init=False)
@@ -2216,11 +2218,9 @@ TextArea {
 
     def action_copy(self) -> None:
         """Copy selection to clipboard."""
-        start, end = self.selection
-        if start == end:
-            return
-        copy_text = self.get_text_range(start, end)
-        self.app.copy_to_clipboard(copy_text)
+        selected_text = self.selected_text
+        if selected_text:
+            self.app.copy_to_clipboard(selected_text)
 
     def action_paste(self) -> None:
         """Paste from local clipboard."""
@@ -2310,10 +2310,6 @@ TextArea {
             to_location = (cursor_row, current_row_length)
 
         self._delete_via_keyboard(end, to_location)
-
-    def action_copy_selection(self) -> None:
-        """Copy the current selection to the clipboard."""
-        self.app.copy_to_clipboard(self.selected_text)
 
 
 @lru_cache(maxsize=128)